refactor: move form into a separate component file.
