+++ src/app/form/[slug]/attachments/[id]/page.tsx
@@ -1,8 +1,7 @@
 'use client'
 
-import React, { useState } from "react";
-import { useFormStatus } from "react-dom";
-import { useRouter } from "next/navigation";
+import { useState } from "react";
+import Form from '@/app/components/receipts/form/Form';
 
 const getReceiptData = () => {
     return {
@@ -20,61 +19,13 @@ const Title = ({ text } : { text: string }) => {
 }
 
 const Upload = () => {
-    const router = useRouter();
-    const [receiptData, setReceiptData] = useState(getReceiptData());
-    const { pending } = useFormStatus();
-
-    // retrieve from S3 bucket
-    const name = "hotel_for_jim.jpg";
-
-    // will need to become a server function for async/await
-    const handleFormSubmission = async (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        const formData = new FormData(event.currentTarget);
-        const data = {
-            total: parseFloat(formData.get("total") as string) || 0,
-            items: receiptData.items.map((_, index) => ({
-                name: formData.get(`items[${index}].name`),
-                quantity: parseInt(formData.get(`items[${index}].quantity`) as string) || 0,
-                cost: parseFloat(formData.get(`items[${index}].cost`) as string) || 0,
-            }))
-        }
-
-        console.log(data)
-        const slug = 0;     // slug to be sent from attachments
-        router.push(`/form/${slug}/attachments`)
-    }
+    const [receiptData] = useState(getReceiptData());
+    const name = "hotel_for_jim.jpg";   // retrieve image name from S3 bucket
 
     return (
         <>
             <Title text={ name } />
-            <form onSubmit={(event) => handleFormSubmission(event)}>
-                <table>
-                    <tbody>
-                        <tr>
-                            <td colSpan={2} className="text-right">
-                                Total £
-                            </td>
-                            <td><input name="total" type="number" defaultValue={ receiptData ? receiptData.total : 0.00 } />   </td>
-                        </tr>
-                        <tr>
-                            <td>Item Name</td>
-                            <td>No.</td>
-                            <td>Cost £</td>
-                        </tr>
-                        {receiptData && receiptData.items && receiptData.items.map(({ name, quantity, cost}, index) => (
-                            <tr key={ index }>
-                                    <td><input name={`items[${index}].name`} type="text" defaultValue={ name ? name : "" } /></td>
-                                    <td><input name={`items[${index}].quantity`} type="number" defaultValue={ quantity ? quantity : 0 } min="0" /></td>
-                                    <td><input name={`items[${index}].cost`} type="number" defaultValue={ cost ? cost : 0.00 } min="0.00" step="0.01" /></td>
-                            </tr>
-                        ))}
-                    </tbody>
-                </table>
-                <button type="submit" className="hover:cursor-pointer">
-                    { !(pending) ? "Submit" : "Submitting" }
-                </button>
-            </form>
+            <Form receiptData={ receiptData } />
         </>
     )
 }
